+++ .eslintrc.js
@@ -1,10 +1,24 @@
+/* global module */
+
 module.exports = {
-    "parser": "babel-eslint",
-    "extends": "airbnb",
-    "plugins": [
-        "react"
-    ],
-    "env": {
-      "browser": true
-    }
+  env: {
+    browser: true,
+    es2021: true,
+  },
+  parser: "@typescript-eslint/parser",
+  plugins: ["@typescript-eslint"],
+  extends: [
+    "eslint:recommended",
+    "plugin:@typescript-eslint/recommended",
+    "plugin:react/recommended",
+    "plugin:react-hooks/recommended",
+  ],
+  parserOptions: {
+    ecmaVersion: 12,
+    sourceType: "module",
+  },
+  rules: {
+    "react/jsx-uses-react": "off",
+    "react/react-in-jsx-scope": "off",
+  },
 };